career-advisor/backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
openai==1.102.0
python-decouple==3.8
scikit-learn>=1.3.0,<1.6.0
numpy>=1.24.0,<2.0
pandas>=2.0,<2.3
joblib==1.4.2